gemini-2.5-app/requirements.txt
fastapi
uvicorn
gradio
google-genai
python-dotenv
